feat: sprinkles 수정
