+++ packages/ui/src/globalStyle/theme.css.ts
@@ -102,13 +102,3 @@ export const Color = createGlobalTheme(":root", {
 });
 
 export type Colors = FlattenObjectKeys<typeof Color>;
-
-//반응형 미디어쿼리
-export const Media = createGlobalTheme(":root", {
-  SCREEN_SM: "screen and (min-width: 480px)",
-  SCREEN_MD: "screen and (min-width: 768px)",
-  SCREEN_LG: "screen and (min-width: 1024px)",
-  SCREEN_XL: "screen and (min-width: 1280px)",
-});
-
-export type MEDIA = typeof Media;
